// server/src/tree-sitter-javascript.d.ts
interface SyntaxNodeBase {
  tree: Tree;
  type: string;
  isNamed: boolean;
  text: string;
  startPosition: Point;
  endPosition: Point;
  startIndex: number;
  endIndex: number;
  parent: SyntaxNode | null;
  children: Array<SyntaxNode>;
  namedChildren: Array<SyntaxNode>;
  childCount: number;
  namedChildCount: number;
  firstChild: SyntaxNode | null;
  firstNamedChild: SyntaxNode | null;
  lastChild: SyntaxNode | null;
  lastNamedChild: SyntaxNode | null;
  nextSibling: SyntaxNode | null;
  nextNamedSibling: SyntaxNode | null;
  previousSibling: SyntaxNode | null;
  previousNamedSibling: SyntaxNode | null;

  hasChanges(): boolean;
  hasError(): boolean;
  isMissing(): boolean;
  toString(): string;
  child(index: number): SyntaxNode | null;
  namedChild(index: number): SyntaxNode | null;
  firstChildForIndex(index: number): SyntaxNode | null;
  firstNamedChildForIndex(index: number): SyntaxNode | null;

  descendantForIndex(index: number): SyntaxNode;
  descendantForIndex(startIndex: number, endIndex: number): SyntaxNode;
  namedDescendantForIndex(index: number): SyntaxNode;
  namedDescendantForIndex(startIndex: number, endIndex: number): SyntaxNode;
  descendantForPosition(position: Point): SyntaxNode;
  descendantForPosition(startPosition: Point, endPosition: Point): SyntaxNode;
  namedDescendantForPosition(position: Point): SyntaxNode;
  namedDescendantForPosition(startPosition: Point, endPosition: Point): SyntaxNode;
  descendantsOfType<T extends TypeString>(types: T | readonly T[], startPosition?: Point, endPosition?: Point): NodeOfType<T>[];

  closest<T extends SyntaxType>(types: T | readonly T[]): NamedNode<T> | null;
  walk(): TreeCursor;
}
// ...
interface NamedNodeBase extends SyntaxNodeBase {
    isNamed: true;
}
// ...
export type NodeOfType<T extends string> = PickType<SyntaxNode, T>;

interface TreeCursorOfType<S extends string, T extends SyntaxNodeBase> {
  nodeType: S;
  currentNode: T;
}

type TreeCursorRecord = { [K in TypeString]: TreeCursorOfType<K, NodeOfType<K>> };
// ...
export type TypedTreeCursor = TreeCursorRecord[keyof TreeCursorRecord];

export interface ErrorNode extends NamedNodeBase {
    type: SyntaxType.ERROR;
    hasError(): true;
}

export const enum SyntaxType {
  ERROR = "ERROR",
  Arguments = "arguments",
  Array = "array",
  ArrayPattern = "array_pattern",
  ArrowFunction = "arrow_function",
  AssignmentExpression = "assignment_expression",
  AssignmentPattern = "assignment_pattern",
  AugmentedAssignmentExpression = "augmented_assignment_expression",
  AwaitExpression = "await_expression",
  BinaryExpression = "binary_expression",
  BreakStatement = "break_statement",
  CallExpression = "call_expression",
  CatchClause = "catch_clause",
  Class = "class",
  ClassBody = "class_body",
  ClassDeclaration = "class_declaration",
  ClassHeritage = "class_heritage",
  ComputedPropertyName = "computed_property_name",
  ContinueStatement = "continue_statement",
  DebuggerStatement = "debugger_statement",
  Decorator = "decorator",
  DoStatement = "do_statement",
  EmptyStatement = "empty_statement",
  ExportClause = "export_clause",
  ExportSpecifier = "export_specifier",
  ExportStatement = "export_statement",
  ExpressionStatement = "expression_statement",
  FinallyClause = "finally_clause",
  ForInStatement = "for_in_statement",
  ForStatement = "for_statement",
  FormalParameters = "formal_parameters",
  Function = "function",
  FunctionDeclaration = "function_declaration",
  GeneratorFunction = "generator_function",
  GeneratorFunctionDeclaration = "generator_function_declaration",
  IfStatement = "if_statement",
  Import = "import",
  ImportClause = "import_clause",
  ImportSpecifier = "import_specifier",
  ImportStatement = "import_statement",
  JsxAttribute = "jsx_attribute",
  JsxClosingElement = "jsx_closing_element",
  JsxElement = "jsx_element",
  JsxExpression = "jsx_expression",
  JsxFragment = "jsx_fragment",
  JsxNamespaceName = "jsx_namespace_name",
  JsxOpeningElement = "jsx_opening_element",
  JsxSelfClosingElement = "jsx_self_closing_element",
  LabeledStatement = "labeled_statement",
  LexicalDeclaration = "lexical_declaration",
  MemberExpression = "member_expression",
  MetaProperty = "meta_property",
  MethodDefinition = "method_definition",
  NamedImports = "named_imports",
  NamespaceImport = "namespace_import",
  NestedIdentifier = "nested_identifier",
  NewExpression = "new_expression",
  Object = "object",
  ObjectPattern = "object_pattern",
  Pair = "pair",
  ParenthesizedExpression = "parenthesized_expression",
  Program = "program",
  PublicFieldDefinition = "public_field_definition",
  Regex = "regex",
  RestParameter = "rest_parameter",
  ReturnStatement = "return_statement",
  SequenceExpression = "sequence_expression",
  SpreadElement = "spread_element",
  StatementBlock = "statement_block",
  String = "string",
  SubscriptExpression = "subscript_expression",
  SwitchBody = "switch_body",
  SwitchCase = "switch_case",
  SwitchDefault = "switch_default",
  SwitchStatement = "switch_statement",
  TemplateString = "template_string",
  TemplateSubstitution = "template_substitution",
  TernaryExpression = "ternary_expression",
  ThrowStatement = "throw_statement",
  TryStatement = "try_statement",
  UnaryExpression = "unary_expression",
  UpdateExpression = "update_expression",
  VariableDeclaration = "variable_declaration",
  VariableDeclarator = "variable_declarator",
  WhileStatement = "while_statement",
  WithStatement = "with_statement",
  YieldExpression = "yield_expression",
  EscapeSequence = "escape_sequence",
  False = "false",
  HashBangLine = "hash_bang_line",
  Identifier = "identifier",
  JsxText = "jsx_text",
  Null = "null",
  Number = "number",
  PropertyIdentifier = "property_identifier",
  RegexFlags = "regex_flags",
  RegexPattern = "regex_pattern",
  ShorthandPropertyIdentifier = "shorthand_property_identifier",
  StatementIdentifier = "statement_identifier",
  Super = "super",
  This = "this",
  True = "true",
  Undefined = "undefined",
}
// ...
export type TypeString = SyntaxType | UnnamedType;
// ...
export interface ClassBodyNode extends NamedNodeBase {
  type: SyntaxType.ClassBody;
  memberNodes: (MethodDefinitionNode | PublicFieldDefinitionNode)[];
}

export interface ClassDeclarationNode extends NamedNodeBase {
  type: SyntaxType.ClassDeclaration;
  bodyNode: ClassBodyNode;
  decoratorNodes: DecoratorNode[];
  nameNode: IdentifierNode;
}
// ...
export interface DecoratorNode extends NamedNodeBase {
  type: SyntaxType.Decorator;
}
// ...
export interface MethodDefinitionNode extends NamedNodeBase {
  type: SyntaxType.MethodDefinition;
  bodyNode: StatementBlockNode;
  decoratorNodes: DecoratorNode[];
  nameNode: ComputedPropertyNameNode | NumberNode | PropertyIdentifierNode | StringNode;
  parametersNode: FormalParametersNode;
}
// ...
export interface PublicFieldDefinitionNode extends NamedNodeBase {
  type: SyntaxType.PublicFieldDefinition;
  propertyNode: ComputedPropertyNameNode | NumberNode | PropertyIdentifierNode | StringNode;
  valueNode?: ExpressionNode;
}
// ...
export interface IdentifierNode extends NamedNodeBase {
  type: SyntaxType.Identifier;
}